[FEAT] 주문 취소 및 수정 기능 구현 (#35)
* [feat] 주문 취소 구현

* [feat] 주문 수정 구현

* [fix] 주문 수량 자료형 변경 (BigInteger -> Long)

+++ order-service/src/main/java/com/vroomvroom/orderservice/infrastructure/dto/ProductDTO.java
@@ -22,5 +22,5 @@ import java.util.UUID;
 public class ProductDTO {
     private UUID productId;
     private Money price;
-    private BigInteger stock;
+    private Long stock;
 }
